+++ protocols/kad/CHANGELOG.md
@@ -6,9 +6,12 @@
 
 - Remove lifetime from `RecordStore` and use GATs instead. See [PR 3239].
 
+- Limit number of active outbound streams to 32. See [PR 3287].
+
 - Bump MSRV to 1.65.0.
 
 [PR 3239]: https://github.com/libp2p/rust-libp2p/pull/3239
+[PR 3287]: https://github.com/libp2p/rust-libp2p/pull/3287
 
 # 0.42.1